# web-service/src/main/resources/web-service-openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:63294
  description: Generated server url
paths:
  /api/v1/orders/{id}:
    get:
      tags:
      - order-controller
      operationId: getById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderEntity"
    put:
      tags:
      - order-controller
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderEntity"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderEntity"
    delete:
      tags:
      - order-controller
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/orders:
    post:
      tags:
      - order-controller
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderEntity"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderEntity"
  /api/v1/ping:
    get:
      tags:
      - test-controller
      operationId: ping
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
components:
  schemas:
    Allocation:
      type: object
      properties:
        id:
          type: string
        quantity:
          type: integer
          format: int32
      required:
      - id
      - quantity
    Limit:
      allOf:
      - $ref: "#/components/schemas/OrderType"
      - type: object
        properties:
          price:
            type: number
      required:
      - price
    LimitOnClose:
      allOf:
      - $ref: "#/components/schemas/OrderType"
      - type: object
        properties:
          price:
            type: number
      required:
      - price
    Market:
      allOf:
      - $ref: "#/components/schemas/OrderType"
    MarketOnClose:
      allOf:
      - $ref: "#/components/schemas/OrderType"
    OrderDetails:
      type: object
      properties:
        allocations:
          type: array
          items:
            $ref: "#/components/schemas/Allocation"
        orderType:
          oneOf:
          - $ref: "#/components/schemas/Limit"
          - $ref: "#/components/schemas/LimitOnClose"
          - $ref: "#/components/schemas/Market"
          - $ref: "#/components/schemas/MarketOnClose"
          - $ref: "#/components/schemas/StopLimit"
          - $ref: "#/components/schemas/StopLoss"
      required:
      - allocations
      - orderType
    OrderEntity:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderDetails:
          $ref: "#/components/schemas/OrderDetails"
      required:
      - orderDetails
    OrderType:
      discriminator:
        propertyName: type
      properties:
        type:
          type: string
      required:
      - type
    StopLimit:
      allOf:
      - $ref: "#/components/schemas/OrderType"
      - type: object
        properties:
          stopPrice:
            type: number
          limitPrice:
            type: number
      required:
      - limitPrice
      - stopPrice
    StopLoss:
      allOf:
      - $ref: "#/components/schemas/OrderType"
      - type: object
        properties:
          stopPrice:
            type: number
      required:
      - stopPrice